% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_stats.R
\name{analysis_summary_table}
\alias{analysis_summary_table}
\title{Function to create descriptive statistics table with gtsummary}
\usage{
analysis_summary_table(
  data,
  group_col = "group",
  tissue_col = "tissue",
  cols = NULL
)
}
\arguments{
\item{data}{dataframe to be analyzed}

\item{group_col}{contains grouping variable}

\item{tissue_col}{if the dataframe contains multiple rows from the same animal (such as different tissues) this will subset the data}

\item{cols}{specify columns to be analyzed. if blank all numeric columns are analyzed}
}
\description{
Function to create descriptive statistics table with gtsummary
}
