% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{create_posthoc_tables}
\alias{create_posthoc_tables}
\title{Function to create summary table of post-hoc tests. Performs Kruskal-Wallis on all numeric columns and post-hoc Dunn's on all significant columns.}
\usage{
create_posthoc_tables(data, group_col = "group", tissue_col = "tissue")
}
\arguments{
\item{data}{dataframe to be analyzed}

\item{group_col}{contains grouping variable}

\item{tissue_col}{if the dataframe contains multiple rows from the same animal (such as different tissues) this will subset the data}
}
\description{
Function to create summary table of post-hoc tests. Performs Kruskal-Wallis on all numeric columns and post-hoc Dunn's on all significant columns.
}
