% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{import_fcs}
\alias{import_fcs}
\title{Function to automate import of FlowJo workspace data automatically and work with lists. Enables clean import from .wsp with
simultaneous writing to excel.}
\usage{
import_fcs(clean, path, group, r_stats, keywords)
}
\arguments{
\item{clean}{logic, decide if data should be newly imported from .wsp and then saved to excel or
simply from previously created Excel}

\item{path}{location of wsp}

\item{group}{name of group in which samples are organized}

\item{r_stats}{logic indicating if stats (MFI etc.) should be imported}

\item{keywords}{character vector of keywords to be imported}
}
\description{
Function to automate import of FlowJo workspace data automatically and work with lists. Enables clean import from .wsp with
simultaneous writing to excel.
}
