% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{import_fcs_clean}
\alias{import_fcs_clean}
\title{Import and cache FlowJo workspace data to Excel}
\usage{
import_fcs_clean(clean, location, wsp, group, r_stats, keywords)
}
\arguments{
\item{clean}{logic, decide if data should be newly imported from .wsp and then saved to excel or
simply from previously created Excel}

\item{location}{name of subdirectory where data is located}

\item{wsp}{name of workspace file}

\item{group}{name of group in which samples are organized}

\item{r_stats}{logic indicating if stats (MFI etc.) should be imported}

\item{keywords}{character vector of keywords to be imported}
}
\description{
Import and cache FlowJo workspace data to Excel
}
