% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{import_workspace}
\alias{import_workspace}
\title{Import FlowJo workspace data using fcexpr package}
\usage{
import_workspace(path, group, r_stats, keywords)
}
\arguments{
\item{path}{path to FlowJo Space}

\item{group}{group of samples from workspace to include in analysis}

\item{r_stats}{import stats such as MFIs?}

\item{keywords}{which keywords to export}
}
\description{
Import FlowJo workspace data using fcexpr package
}
