% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{import_workspace_long}
\alias{import_workspace_long}
\title{Import FACS data from FlowJo workspace in long format}
\usage{
import_workspace_long(path, group = NULL, r_stats = FALSE, keywords = NULL)
}
\arguments{
\item{path}{path to .wsp file}

\item{group}{group to be extracted from workspace default = NULL}

\item{r_stats}{true or false to indicate if statistics such as MFI should be exported}

\item{keywords}{character vector of keywords to be exported, default = NULL}
}
\description{
Import FACS data from FlowJo workspace in long format
}
