% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_functions.R
\name{knit_exp_structure}
\alias{knit_exp_structure}
\title{New function for automatic knitting of dated files in subdirectory based on new experiment based subdirectories}
\usage{
knit_exp_structure(input, format, ...)
}
\arguments{
\item{input}{unchanged}

\item{format}{specify desired files to output in quotes: "html_document", "pdf_document", "all"}

\item{...}{further arguments}
}
\description{
New function for automatic knitting of dated files in subdirectory based on new experiment based subdirectories
}
