% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_functions.R
\name{knit_multiple_dated}
\alias{knit_multiple_dated}
\title{Function for automatic knitting of dated files in subdirectory}
\usage{
knit_multiple_dated(input, path, ...)
}
\arguments{
\item{input}{unchanged}

\item{path}{change to name of desired subdirectory}

\item{...}{further arguments}
}
\description{
Function for automatic knitting of dated files in subdirectory
}
