% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_functions.R
\name{knit_panel}
\alias{knit_panel}
\title{Panel knitting}
\usage{
knit_panel(input, path, format, ...)
}
\arguments{
\item{input}{unchanged}

\item{path}{path of experiment}

\item{format}{pdf_document}

\item{...}{further parameters}
}
\description{
create panel
}
