% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_knit.R
\name{report_knit_dated}
\alias{report_knit_dated}
\title{Function for automatic knitting of dated files in subdirectory}
\usage{
report_knit_dated(input, path, format, ...)
}
\arguments{
\item{input}{unchanged}

\item{path}{change to name of desired subdirectory}

\item{format}{specify desired files to output in quotes: "html_document", "pdf_document", "all"}

\item{...}{further arguments}
}
\description{
Function for automatic knitting of dated files in subdirectory
}
