% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_knit.R
\name{report_knit_exp}
\alias{report_knit_exp}
\title{New function for automatic knitting of dated files in subdirectory based on new experiment based subdirectories}
\usage{
report_knit_exp(input, format, ...)
}
\arguments{
\item{input}{unchanged}

\item{format}{specify desired files to output in quotes: "html_document", "pdf_document", "all"}

\item{...}{further arguments}
}
\value{
path to the rendered output file, returned by \code{rmarkdown::render()}
}
\description{
New function for automatic knitting of dated files in subdirectory based on new experiment based subdirectories
}
\examples{
\dontrun{
  report_knit_exp(
    input  = "scripts/flow-report.Rmd",
    format = "html_document"
  )
}
}
