% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_knit.R
\name{report_knit_wide}
\alias{report_knit_wide}
\title{function to generate wide html}
\usage{
report_knit_wide(input, ...)
}
\arguments{
\item{input}{file to be knitted}

\item{...}{further arguments}
}
\value{
path to the rendered output file, returned by \code{rmarkdown::render()}
}
\description{
function to generate wide html
}
\examples{
\dontrun{
  report_knit_wide(input = "scripts/flow-report.Rmd")
}
}
