% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{update_JanisHelpers}
\alias{update_JanisHelpers}
\title{update_JanisHelpers: automatically update the package}
\usage{
update_JanisHelpers()
}
\description{
update_JanisHelpers: automatically update the package
}
