% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{update_JanisHelpers_git}
\alias{update_JanisHelpers_git}
\title{update_JanisHelpers_git: update from private git repo}
\usage{
update_JanisHelpers_git()
}
\description{
update_JanisHelpers_git: update from private git repo
}
