% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-functions.R
\name{import_fcs_clean}
\alias{import_fcs_clean}
\title{Function to automate import of FlowJo workspace data automatically. Enables clean import from .wsp with
simultaneous writing to excel.}
\usage{
import_fcs_clean(clean, location, wsp, group, stats, keywords)
}
\arguments{
\item{clean}{logic, decide if data should be newly imported from .wsp and then saved to excel or
simply from previously created Excel}

\item{location}{name of subdirectory where data is located}

\item{wsp}{name of workspace file}

\item{group}{name of group in which samples are organized}

\item{stats}{logic indicating if stats should be imported}

\item{keywords}{character vector of keywords to be imported}
}
\description{
Function to automate import of FlowJo workspace data automatically. Enables clean import from .wsp with
simultaneous writing to excel.
}
