% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_functions.R
\name{knit_wide_html}
\alias{knit_wide_html}
\title{function to generate wide html}
\usage{
knit_wide_html(input, ...)
}
\arguments{
\item{input}{file to be knitted}

\item{...}{further arguments}

\item{format}{defaults to wide html_document}
}
\description{
function to generate wide html
}
