% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_functions.R
\name{knit_with_date}
\alias{knit_with_date}
\title{Knits multiple files into subdirectory and adds date to filename}
\usage{
knit_with_date(input, ...)
}
\description{
Knits multiple files into subdirectory and adds date to filename
}
